Fixed int rounding in PWM setting

+++ pwm_manager.c
@@ -24,7 +24,7 @@ void init_pwm_pin(uint pin) {
  * \param throttle throttle value [0 full reverse, 127 full stop, 254 full forward]
  */
 void set_pwm_pin(uint pin, char throttle) {
-    int duty_cycle;
+    float duty_cycle;
     if (throttle == 127) {
         duty_cycle = get_duty_cycle(PWM_FREQ, FULL_STOP_MS);
     } else {
@@ -38,13 +38,13 @@ void set_pwm_pin(uint pin, char throttle) {
 }
 
 
-static int get_duty_cycle(uint freq, float pulse_len_ms) {
-    float total_pulse_ms = 1.0/freq*1000;
-    return (pulse_len_ms/total_pulse_ms)*100;
+static float get_duty_cycle(uint freq, float pulse_len_ms) {
+    float total_pulse_ms = 1.0/freq*1000.0;
+    return (pulse_len_ms/total_pulse_ms)*100.0;
 }
 
 
-static uint32_t pwm_set_freq_duty(uint slice_num, uint chan, uint32_t freq, int duty_cycle) {
+static uint32_t pwm_set_freq_duty(uint slice_num, uint chan, uint32_t freq, float duty_cycle) {
     uint32_t clock = 125000000;
     uint32_t divider16 = clock / freq / 4096 + (clock % (freq * 4096) != 0);
     if (divider16 / 16 == 0) divider16 = 16;
@@ -53,7 +53,7 @@ static uint32_t pwm_set_freq_duty(uint slice_num, uint chan, uint32_t freq, int 
 
     pwm_set_clkdiv_int_frac4(slice_num, divider16/16, divider16 & 0xF);
     pwm_set_wrap(slice_num, wrap);
-    pwm_set_chan_level(slice_num, chan, wrap * duty_cycle / 100);
+    pwm_set_chan_level(slice_num, chan, (int)(wrap * duty_cycle / 100.0));
 
     return wrap;
 }
