+++ pwm_manager.h
@@ -18,8 +18,8 @@
 void init_pwm_pin(uint pin);
 void set_pwm_pin(uint pin, char throttle);
 
-static uint32_t pwm_set_freq_duty(uint slice_num, uint chan, uint32_t freq, int duty_cycle);
-static int get_duty_cycle(uint freq, float pulse_len_ms);
+static uint32_t pwm_set_freq_duty(uint slice_num, uint chan, uint32_t freq, float duty_cycle);
+static float get_duty_cycle(uint freq, float pulse_len_ms);
 static float naive_lerp(float a, float b, float t);
 
 #endif
